Reduce server list calls
While running ssh_config, apply_resources was running nova server list once
for every nodes in that build, which is bit costly as it is calling neutron
port list which is causing issues to contrail as of now.

This patch is to reduce the calls, by making get_ip_of_node to accept a list of
nodes and run server list once and return all server's ip.